Version 0.7
Includes VimLeavePre hook to dump the contents of the scratch buffer to a file before exiting.  Set the global variable g:scratchBackupFile to a file path to enable the feature. See comments in the plugin file.

(Once again , Thanks to Hari Krishna Dara (hari_vim at yahoo dot com) for the useful enhancement!)